+++ test-cases/syntax-1-instanced-fn/webpack.config.maker.js
@@ -18,12 +18,13 @@ function webpackConfig(index, mode) {
           use: 'babel-loader',
         },
         {
-          test: /\.css$/,
+          test: /\.scss$/,
           use: [
             index === 1 ? MiniCssExtractPlugin.loader : 'style-loader',
             {
               loader: 'css-loader',
               options: {
+                importLoaders: 2,
                 modules: {
                   exportLocalsConvention: 'camelCase',
                   getLocalIdent: minifyCssIdents.getLocalIdent,
@@ -34,6 +35,14 @@ function webpackConfig(index, mode) {
               },
             },
             'postcss-loader',
+            {
+              loader: 'sass-loader',
+              options: {
+                sassOptions: {
+                  indentWidth: 2,
+                },
+              },
+            },
           ],
         },
       ],