+++ test-cases/syntax-4-loader-with-plugin/webpack.config.maker.js
@@ -17,12 +17,13 @@ function webpackConfig(index, mode) {
           use: 'babel-loader',
         },
         {
-          test: /\.css$/,
+          test: /\.scss$/,
           use: [
             index === 1 ? MiniCssExtractPlugin.loader : 'style-loader',
             {
               loader: 'minify-css-idents/css-loader',
               options: {
+                importLoaders: 2,
                 modules: {
                   exportLocalsConvention: 'camelCase',
                   localIdentContext,
@@ -32,6 +33,14 @@ function webpackConfig(index, mode) {
               },
             },
             'postcss-loader',
+            {
+              loader: 'sass-loader',
+              options: {
+                sassOptions: {
+                  indentWidth: 2,
+                },
+              },
+            },
           ],
         },
       ],