+++ test-cases/syntax-5-loader-without-plugin/webpack.config.maker.js
@@ -17,12 +17,13 @@ function webpackConfig(index, mode) {
           use: 'babel-loader',
         },
         {
-          test: /\.css$/,
+          test: /\.scss$/,
           use: [
             MiniCssExtractPlugin.loader,
             {
               loader: 'minify-css-idents/css-loader',
               options: {
+                importLoaders: 2,
                 modules: {
                   exportLocalsConvention: 'camelCase',
                   localIdentName: '[path]___[name]__[local]',
@@ -31,6 +32,14 @@ function webpackConfig(index, mode) {
               },
             },
             'postcss-loader',
+            {
+              loader: 'sass-loader',
+              options: {
+                sassOptions: {
+                  indentWidth: 2,
+                },
+              },
+            },
           ],
         },
       ],